p2p/discover: merge topic eviction tests into v5_topic_test.go
